HomeKit Fixup
Improvements to HomeKit associating methods, e.g. method re-use. 

In HAP:
the methods for reading/writing the key stores are now abstracted away.
The classes, reusable. Reduce, Reuse, Recycle... OK, two out of three.
The main point of this commit is to fix file-handling for unpredictable
device ID names. Usually they're like a MAC address, but future spec 
might say something else. So we store those in a JSON file now.

This commit also removes the dependency on NaCl lib, which resolves 
issue #24. Until such a time as when nacl is really necessary, and not
a stop-gap for the lazy developers :)

Fleshed out many HAP methods with accompanying comments from the HAP
spec doc, which outline what is not yet implemented, and help ease you
into the crypto side of what's going on.