RTPBuffered Audio improvements. Less snap crackle and pop. Smooth. -RTPBuffered scenarios seek instantly. -No more old loop buffer - she has served us well. -The RTPBuffered improvements also won near-perfect audio sync.
-Rewrite of real-time audio buffer. No synchronization with others yet.
That seems to need the most amount of work.

"The right way" seems to be to decode audio into larger blocks,
buffer that, and compensating for output device latency, play it.
Real-time frames are ALAC, 352 samples long, which is ~8 msec.
Anything longer than 8 msec to process the next output frame can cause
hiccups or skips as we deal with audio packet-wise, with a just in time
approach. Problems seem to start in PulseAudio, which both pyaudio
and sounddevice seem to rely on. Sometimes processing a frame takes
longer than 8 msec, because... reasons. Increasing the device buffer
lowers CPU usage to improve things somewhat. I avg about 20-40 µsec per
ALAC frame.

One could amortize against the occasional CPU blip with this surplus of
free time via a double-buffer and likely achieve stable output.
Optionally via a separate process. Alternatively, resample each frame to
be slightly longer or shorter to compensate. 

Tried:
-callbacks but that presents its own challenges.
-changing Pyaudio to sounddevice, same results.

So for now, leave real-time.